+++ cookie.go
@@ -5,26 +5,48 @@ import (
 	"time"
 )
 
-// Config is the set of cookie properties.
-type Config struct {
-	// cookie domain/path scope (leave zeroed for requested resource scope)
-	Path   string
+// DefaultCookieConfig configures http.Cookie creation for production.
+var DefaultCookieConfig = &CookieConfig{
+	Path:     "/",
+	MaxAge:   defaultMaxAge,
+	HTTPOnly: true,
+	Secure:   true,
+	SameSite: http.SameSiteStrictMode,
+}
+
+// DebugCookieConfig configures http.Cookie creation for debugging. It
+// does NOT require cookies be sent over HTTPS so it should only be used
+// in development. Prefer DefaultCookieConfig.
+var DebugCookieConfig = &CookieConfig{
+	Path:     "/",
+	MaxAge:   defaultMaxAge,
+	HTTPOnly: true,
+	Secure:   false,
+	SameSite: http.SameSiteLaxMode,
+}
+
+// CookieConfig configures http.Cookie creation.
+type CookieConfig struct {
+	// Cookie domain/path scope (leave zeroed for requested resource scope)
+	// Defaults to the domain name of the responding server when unset
 	Domain string
+	// Defaults to the path of the responding URL when unset
+	Path string
 	// MaxAge=0 means no 'Max-Age' attribute specified.
 	// MaxAge<0 means delete cookie now, equivalently 'Max-Age: 0'.
 	// MaxAge>0 means Max-Age attribute present and given in seconds.
 	MaxAge int
-	// cookie may only be transferred over HTTPS
+	// cookie may only be transferred over HTTPS. Recommend true.
 	Secure bool
-	// browser should prohibit non-HTTP (i.e. javascript) cookie access
+	// browser should prohibit non-HTTP (i.e. javascript) cookie access. Recommend true
 	HTTPOnly bool
-	// prohibit sending in cross-site requests with SameSiteLaxMode or SameSiteLaxMode
+	// prohibit sending in cross-site requests with SameSiteLaxMode or SameSiteStrictMode
 	SameSite http.SameSite
 }
 
 // newCookie returns a new http.Cookie with the given name, value, and
 // properties from config.
-func newCookie(name, value string, config *Config) *http.Cookie {
+func newCookie(name, value string, config *CookieConfig) *http.Cookie {
 	cookie := &http.Cookie{
 		Name:     name,
 		Value:    value,